Improve logic for saving data
