+++ data/authors.json
@@ -6,5 +6,21 @@
   {
     "first_name": "Stphen ",
     "last_name": "KIng"
+  },
+  {
+    "first_name": "bruh ",
+    "last_name": "what"
+  },
+  {
+    "first_name": "bruh",
+    "last_name": "me"
+  },
+  {
+    "first_name": "ME",
+    "last_name": "ME"
+  },
+  {
+    "first_name": "First",
+    "last_name": "last"
   }
 ]
